Fix `NoDataToSpill` when multiple storage quota requests happen simultaneously (#2203)
(cherry picked from commit 717d8cc0835ad032f406e594a523d91fd203640f)
