+++ mars/storage/core.py
@@ -16,7 +16,7 @@
 
 import asyncio
 from abc import ABC, abstractmethod
-from concurrent.futures import Executor
+from concurrent.futures import Executor, ThreadPoolExecutor
 from typing import Any, Optional, Union
 
 from ..lib.aio import AioFileObject
@@ -29,6 +29,8 @@ class StorageFileObject(AioFileObject):
                  loop: asyncio.BaseEventLoop = None,
                  executor: Executor = None):
         self._object_id = object_id
+        if executor is None:
+            executor = ThreadPoolExecutor()
         super().__init__(file, loop=loop, executor=executor)
 
     @property
